requirements.txt
Django==4.2.10
psycopg2-binary==2.9.9
Pillow==10.2.0
django-allauth==0.61.1
djangorestframework==3.14.0
django-filter==23.5
python-dotenv==1.0.1
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-environ==0.11.2
drf-yasg==1.21.7 